+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientExternalAuthorizationServerTest.kt
@@ -305,7 +305,7 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
             // just pick the first credential in SD-JWT that is available
             val selectedCredential = credentialIdentifierInfos
                 .first { it.supportedCredentialFormat.format == CredentialFormatEnum.DC_SD_JWT }
-            // client will call clientBrowser.openUrlExternally
+
             client.startProvisioningWithAuthRequestReturningResult(
                 credentialIssuerUrl = issuerPublicContext,
                 credentialIdentifierInfo = selectedCredential,