+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientIntegratedDPoPTest.kt
@@ -2,32 +2,26 @@ package at.asitplus.wallet.lib.ktor.openid
 
 import at.asitplus.catching
 import at.asitplus.openid.CredentialFormatEnum
-import at.asitplus.openid.IssuerMetadata
-import at.asitplus.openid.OAuth2AuthorizationServerMetadata
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.TokenRequestParameters
 import at.asitplus.signum.indispensable.josef.toJwsAlgorithm
+import at.asitplus.testballoon.withFixtureGenerator
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithSelfSignedCert
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.IssuerAgent
 import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.agent.RandomSource
-import at.asitplus.wallet.lib.data.ConstantIndex
-import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.ISO_MDOC
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
 import at.asitplus.wallet.lib.data.rfc3986.toUri
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.jws.JwsHeaderCertOrJwk
 import at.asitplus.wallet.lib.jws.JwsHeaderNone
 import at.asitplus.wallet.lib.jws.SignJwt
-import at.asitplus.wallet.lib.ktor.openid.TestUtils.credentialDataProviderFun
-import at.asitplus.wallet.lib.ktor.openid.TestUtils.dummyUser
 import at.asitplus.wallet.lib.ktor.openid.TestUtils.respond
-import at.asitplus.wallet.lib.ktor.openid.TestUtils.respondOAuth2Error
+import at.asitplus.wallet.lib.ktor.openid.TestUtils.respondWithDpopNoncePar
+import at.asitplus.wallet.lib.ktor.openid.TestUtils.respondWithDpopNonceToken
 import at.asitplus.wallet.lib.ktor.openid.TestUtils.toRequestInfo
-import at.asitplus.wallet.lib.ktor.openid.TestUtils.verifyIsoMdocCredential
 import at.asitplus.wallet.lib.ktor.openid.TestUtils.verifySdJwtCredential
 import at.asitplus.wallet.lib.oauth2.ClientAuthenticationService
 import at.asitplus.wallet.lib.oauth2.OAuth2Client
@@ -36,15 +30,14 @@ import at.asitplus.wallet.lib.oauth2.TokenService
 import at.asitplus.wallet.lib.oidvci.BuildClientAttestationJwt
 import at.asitplus.wallet.lib.oidvci.CredentialAuthorizationServiceStrategy
 import at.asitplus.wallet.lib.oidvci.CredentialIssuer
-import at.asitplus.wallet.lib.oidvci.IssuerEncryptionService
 import at.asitplus.wallet.lib.oidvci.WalletService
 import at.asitplus.wallet.lib.oidvci.decodeFromPostBody
 import at.asitplus.wallet.lib.oidvci.decodeFromUrlQuery
 import com.benasher44.uuid.uuid4
 import de.infix.testBalloon.framework.core.testSuite
 import io.github.aakira.napier.Napier
+import io.kotest.assertions.fail
 import io.kotest.matchers.nulls.shouldNotBeNull
-import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.client.*
 import io.ktor.client.engine.mock.*
 import io.ktor.client.request.*
@@ -53,10 +46,13 @@ import io.ktor.util.*
 
 /**
  * Tests [OpenId4VciClient] against [CredentialIssuer] with our own internal [SimpleAuthorizationService].
+ *
+ * Makes sure that the [OpenId4VciClient] and [OAuth2KtorClient] use the DPoP nonce provided in success responses too.
  */
-val OpenId4VciClientWithEncryptionTest by testSuite {
+val OpenId4VciClientIntegratedDPoPTest by testSuite {
 
     data class Context(
+        val attributes: Map<String, String>,
         val credentialKeyMaterial: KeyMaterial,
         val dpopKeyMaterial: KeyMaterial,
         val clientAuthKeyMaterial: KeyMaterial,
@@ -66,15 +62,14 @@ val OpenId4VciClientWithEncryptionTest by testSuite {
         val client: OpenId4VciClient,
     )
 
-    fun setup(
-        scheme: ConstantIndex.CredentialScheme,
-        representation: ConstantIndex.CredentialRepresentation,
-        attributes: Map<String, String>,
-    ): Context {
+    withFixtureGenerator {
+        val scheme = EuPidScheme
+        val representation = SD_JWT
+        val attributes = mapOf(EuPidScheme.Attributes.FAMILY_NAME to uuid4().toString())
         val credentialKeyMaterial = EphemeralKeyWithoutCert()
         val dpopKeyMaterial = EphemeralKeyWithoutCert()
         val clientAuthKeyMaterial = EphemeralKeyWithoutCert()
-        val credentialSchemes = setOf(scheme)
+        val credentialSchemes = setOf<at.asitplus.wallet.lib.data.ConstantIndex.CredentialScheme>(scheme)
         val authorizationEndpointPath = "/authorize"
         val tokenEndpointPath = "/token"
         val credentialEndpointPath = "/credential"
@@ -106,30 +101,21 @@ val OpenId4VciClientWithEncryptionTest by testSuite {
             publicContext = publicContext,
             credentialEndpointPath = credentialEndpointPath,
             nonceEndpointPath = nonceEndpointPath,
-            // that's crucial: require encryption
-            encryptionService = IssuerEncryptionService(
-                requireResponseEncryption = true,
-                requireRequestEncryption = true
-            )
         )
-        val mockEngine = MockEngine.Companion { request ->
+        val mockEngine = MockEngine { request ->
             when {
-                request.url.rawSegments.drop(1) == OpenIdConstants.WellKnownPaths.CredentialIssuer -> respond(
-                    vckJsonSerializer.encodeToString<IssuerMetadata>(credentialIssuer.metadata),
-                    headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                )
+                request.url.rawSegments.drop(1) == OpenIdConstants.WellKnownPaths.CredentialIssuer ->
+                    this.respond(credentialIssuer.metadata)
 
-                request.url.rawSegments.drop(1) == OpenIdConstants.WellKnownPaths.OauthAuthorizationServer -> respond(
-                    vckJsonSerializer.encodeToString<OAuth2AuthorizationServerMetadata>(authorizationService.metadata()),
-                    headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                )
+                request.url.rawSegments.drop(1) == OpenIdConstants.WellKnownPaths.OauthAuthorizationServer ->
+                    this.respond(authorizationService.metadata())
 
                 request.url.fullPath.startsWith(parEndpointPath) -> {
                     val requestBody = request.body.toByteArray().decodeToString()
-                    val authnRequest: RequestParameters = requestBody.decodeFromPostBody()
-                    authorizationService.par(authnRequest, request.toRequestInfo()).fold(
-                        onSuccess = { respond(it) },
-                        onFailure = { respondOAuth2Error(it) }
+                    val authnRequest: RequestParameters = requestBody.decodeFromPostBody<RequestParameters>()
+                    authorizationService.parWithDpopNonce(authnRequest, request.toRequestInfo()).fold(
+                        onSuccess = { this.respondWithDpopNoncePar(it) },
+                        onFailure = { fail("$parEndpointPath should not return an error") }
                     )
                 }
 
@@ -138,25 +124,25 @@ val OpenId4VciClientWithEncryptionTest by testSuite {
                     val queryParameters: Map<String, String> =
                         request.url.parameters.toMap().entries.associate { it.key to it.value.first() }
                     val authnRequest: RequestParameters =
-                        if (requestBody.isEmpty()) queryParameters.decodeFromUrlQuery()
-                        else requestBody.decodeFromPostBody()
-                    authorizationService.authorize(authnRequest) { catching { dummyUser() } }.fold(
-                        onSuccess = { respondRedirect(it.url) },
-                        onFailure = { respondOAuth2Error(it) }
+                        if (requestBody.isEmpty()) queryParameters.decodeFromUrlQuery<RequestParameters>()
+                        else requestBody.decodeFromPostBody<RequestParameters>()
+                    authorizationService.authorize(authnRequest) { this.catching { TestUtils.dummyUser() } }.fold(
+                        onSuccess = { this.respondRedirect(it.url) },
+                        onFailure = { fail("$authorizationEndpointPath should not return an error") }
                     )
                 }
 
                 request.url.fullPath.startsWith(tokenEndpointPath) -> {
                     val requestBody = request.body.toByteArray().decodeToString()
                     val params: TokenRequestParameters = requestBody.decodeFromPostBody<TokenRequestParameters>()
-                    authorizationService.token(params, request.toRequestInfo()).fold(
-                        onSuccess = { respond(it) },
-                        onFailure = { respondOAuth2Error(it) }
+                    authorizationService.tokenWithDpopNonce(params, request.toRequestInfo()).fold(
+                        onSuccess = { this.respondWithDpopNonceToken(it) },
+                        onFailure = { fail("$tokenEndpointPath should not return an error") }
                     )
                 }
 
                 request.url.fullPath.startsWith(nonceEndpointPath) -> {
-                    respond(credentialIssuer.nonceWithDpopNonce().getOrThrow())
+                    this.respond(credentialIssuer.nonceWithDpopNonce().getOrThrow())
                 }
 
                 request.url.fullPath.startsWith(credentialEndpointPath) -> {
@@ -165,21 +151,25 @@ val OpenId4VciClientWithEncryptionTest by testSuite {
                     credentialIssuer.credential(
                         authorizationHeader = authn,
                         params = WalletService.CredentialRequest.parse(requestBody).getOrThrow(),
-                        credentialDataProvider = credentialDataProviderFun(scheme, representation, attributes),
+                        credentialDataProvider = TestUtils.credentialDataProviderFun(
+                            scheme,
+                            representation,
+                            attributes
+                        ),
                         request = request.toRequestInfo(),
                     ).fold(
-                        onSuccess = { respond(it) },
-                        onFailure = { respondOAuth2Error(it) }
+                        onSuccess = { this.respond(it) },
+                        onFailure = { fail("$credentialEndpointPath should not return an error") }
                     )
                 }
 
-                else -> respondError(HttpStatusCode.NotFound)
+                else -> this.respondError(HttpStatusCode.NotFound)
                     .also { Napier.w("NOT MATCHED ${request.url.fullPath}") }
             }
         }
         val clientId = "https://example.com/rp"
-
-        return Context(
+        Context(
+            attributes = attributes,
             credentialKeyMaterial = credentialKeyMaterial,
             dpopKeyMaterial = dpopKeyMaterial,
             clientAuthKeyMaterial = clientAuthKeyMaterial,
@@ -208,67 +198,36 @@ val OpenId4VciClientWithEncryptionTest by testSuite {
                     oAuth2Client = OAuth2Client(clientId = clientId),
                     randomSource = RandomSource.Default,
                 )
-            ),
+            )
         )
-    }
-
-    test("loadEuPidCredentialSdJwt") {
-        val expectedFamilyName = uuid4().toString()
-        val expectedAttributeName = EuPidScheme.Attributes.FAMILY_NAME
-        with(setup(EuPidScheme, SD_JWT, mapOf(expectedAttributeName to expectedFamilyName))) {
+    } - {
+        test("loadEuPidCredentialSdJwt") { context ->
             var refreshTokenStore: RefreshTokenInfo? = null
 
-            // Load credential identifier infos from Issuing service
-            val credentialIdentifierInfos = client.loadCredentialMetadata("http://localhost").getOrThrow()
-            // just pick the first credential in SD-JWT that is available
+            val credentialIdentifierInfos = context.client.loadCredentialMetadata("http://localhost").getOrThrow()
             val selectedCredential = credentialIdentifierInfos
                 .first { it.supportedCredentialFormat.format == CredentialFormatEnum.DC_SD_JWT }
 
-            client.startProvisioningWithAuthRequestReturningResult(
+            context.client.startProvisioningWithAuthRequestReturningResult(
                 credentialIssuerUrl = "http://localhost",
                 credentialIdentifierInfo = selectedCredential,
             ).getOrThrow().also {
                 // Simulates the browser, handling authorization to get the authCode
-                val httpClient = HttpClient(mockEngine) { followRedirects = false }
+                val httpClient = HttpClient(context.mockEngine) { followRedirects = false }
                 val authCode = httpClient.get(it.url).headers[HttpHeaders.Location]
-                client.resumeWithAuthCode(authCode!!, it.context).getOrThrow().also {
+                context.client.resumeWithAuthCode(authCode!!, it.context).getOrThrow().also {
                     refreshTokenStore = it.refreshToken!!
-                    it.verifySdJwtCredential(expectedAttributeName, expectedFamilyName, credentialKeyMaterial.publicKey)
+                    context.attributes.forEach { (key, value) ->
+                        it.verifySdJwtCredential(key, value, context.credentialKeyMaterial.publicKey)
+                    }
                 }
             }
 
             refreshTokenStore.shouldNotBeNull()
-            client.refreshCredentialReturningResult(refreshTokenStore).getOrThrow().also {
-                it.verifySdJwtCredential(expectedAttributeName, expectedFamilyName, credentialKeyMaterial.publicKey)
-            }
-        }
-    }
-
-    test("loadEuPidCredentialIsoWithOffer") {
-        val expectedAttributeValue = uuid4().toString()
-        val expectedAttributeName = EuPidScheme.Attributes.GIVEN_NAME
-        with(setup(EuPidScheme, ISO_MDOC, mapOf(expectedAttributeName to expectedAttributeValue))) {
-            var refreshTokenStore: RefreshTokenInfo? = null
-
-            // Load credential identifier infos from Issuing service
-            val credentialIdentifierInfos = client.loadCredentialMetadata("http://localhost").getOrThrow()
-            // just pick the first credential in MSO_MDOC that is available
-            val selectedCredential = credentialIdentifierInfos
-                .first { it.supportedCredentialFormat.format == CredentialFormatEnum.MSO_MDOC }
-
-            val offer = authorizationService.credentialOfferWithPreAuthnForUser(
-                dummyUser(),
-                credentialIssuer.metadata.credentialIssuer
-            )
-            client.loadCredentialWithOfferReturningResult(offer, selectedCredential, null).getOrThrow().also {
-                it.shouldBeInstanceOf<CredentialIssuanceResult.Success>().also {
-                    refreshTokenStore = it.refreshToken!!
-                    it.verifyIsoMdocCredential(expectedAttributeName, expectedAttributeValue)
+            context.client.refreshCredentialReturningResult(refreshTokenStore).getOrThrow().also {
+                context.attributes.forEach { (key, value) ->
+                    it.verifySdJwtCredential(key, value, context.credentialKeyMaterial.publicKey)
                 }
-            }
-            refreshTokenStore.shouldNotBeNull()
-            client.refreshCredentialReturningResult(refreshTokenStore).getOrThrow().also {
-                it.verifyIsoMdocCredential(expectedAttributeName, expectedAttributeValue)
             }
         }
     }